Add files via upload
This file was created as part of an assignment for my Object Oriented Class. The original objective was to create a program that printed information onto a file and then mindlessly spit the same information onto standard output. That didn't sound like much fun, so I added an i/o feature so that the user can ask to see a specific student's report card. When you run this code, you can either ask for "Bokow, R." , "Fallin, D." , or "Kinglsey, M."